+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/integration/ModelEngine.java
@@ -8,6 +8,8 @@ import org.bukkit.entity.Entity;
 
 public interface ModelEngine {
     boolean applyModel(Entity entity, ConfigurationSection configuration);
+    boolean removeModel(Entity entity, String modelId);
+    boolean removeAllModels(Entity entity);
     boolean removeModelState(Entity entity, @Nullable String modelId, String state, boolean ignoreLerp);
     boolean addModelState(Entity entity, @Nullable String model, String state, int lerpIn, int lerpOut, double speed);
     boolean addSubModel(Entity entity, String modelId, String partId, String subModelId, String subPartId, String customId);